// legacy/app/src/lib/database/types.ts
export type SqliteSynchronous =
	(typeof sqlite_synchronous)[keyof typeof sqlite_synchronous];

export interface CapacitorSQLiteOptions {
	/**
	 * Journal/WAL size limit. Defaults to 6MB.
	 * The WAL may grow larger than this limit during writes, but SQLite will
	 * attempt to truncate the file afterwards.
	 */
	journalSizeLimit?: number;

	/**
	 * SQLite synchronous flag. Defaults to [SqliteSynchronous.normal], which
	 * is safe for WAL mode.
	 */
	synchronous?: SqliteSynchronous;

	/**
	 * Maximum SQLite cache size. Defaults to 50MB.
	 *
	 * For details, see: https://www.sqlite.org/pragma.html#pragma_cache_size
	 */
	cacheSizeKb?: number;
}

/**
 * Object returned by SQL Query executions.
 */
export type QueryResult = {
	/** Represents the auto-generated row id if applicable. */
	insertId?: number;
	/** Number of affected rows if result of a update query. */
	rowsAffected: number;
	/** if status is undefined or 0 this object will contain the query results */
	rows?: {
		/** Raw array with all dataset */
		_array: any[];
		/** The length of the dataset */
		length: number;
		/** A convenience function to acess the index based the row object
		 * @param idx the row index
		 * @returns the row structure identified by column names
		 */
		item: (idx: number) => any;
	};
};

export interface DatabaseGetUtilities {
	/** Execute a read-only query and return results. */
	getAll<T>(sql: string, parameters?: any[]): Promise<T[]>;
	/** Execute a read-only query and return the first result, or null if the ResultSet is empty. */
	getOptional<T>(sql: string, parameters?: any[]): Promise<T | null>;
	/** Execute a read-only query and return the first result, error if the ResultSet is empty. */
	get<T>(sql: string, parameters?: any[]): Promise<T>;
}

export interface LockContext extends DatabaseGetUtilities {
	/** Execute a single write statement. */
	execute: (query: string, params?: any[] | undefined) => Promise<QueryResult>;
	/**
	 * Execute a single write statement and return raw results.
	 * Unlike `execute`, which returns an object with structured key-value pairs,
	 * `executeRaw` returns a nested array of raw values, where each row is
	 * represented as an array of column values without field names.
	 *
	 * Example result:
	 *
	 * ```[ [ '1', 'list 1', '33', 'Post content', '1' ] ]```
	 *
	 * Where as `execute`'s `rows._array` would have been:
	 *
	 * ```[ { id: '33', name: 'list 1', content: 'Post content', list_id: '1' } ]```
	 */
	executeRaw: (query: string, params?: any[] | undefined) => Promise<any[][]>;
}

export interface Transaction extends LockContext {
	/** Commit multiple changes to the local DB using the Transaction context. */
	commit: () => Promise<QueryResult>;
	/** Roll back multiple attempted changes using the Transaction context. */
	rollback: () => Promise<QueryResult>;
}

/**
 * Update table operation numbers from SQLite
 */
export type RowUpdateType =
	(typeof row_update_type)[keyof typeof row_update_type];

export interface TableUpdateOperation {
	opType: RowUpdateType;
	rowId: number;
}
/**
 * Notification of an update to one or more tables, for the purpose of realtime change notifications.
 */
export interface UpdateNotification extends TableUpdateOperation {
	table: string;
}

export interface BatchedUpdateNotification {
	rawUpdates: UpdateNotification[];
	tables: string[];
	groupedUpdates: Record<string, TableUpdateOperation[]>;
}

export interface DBAdapterListener extends BaseListener {
	/**
	 * Listener for table updates.
	 * Allows for single table updates in order to maintain API compatibility
	 * without the need for a major version bump
	 * The DB adapter can also batch update notifications if supported.
	 */
	tablesUpdated: (
		updateNotification: BatchedUpdateNotification | UpdateNotification,
	) => void;
}

export interface DatabaseLockOptions {
	timeoutMs?: number;
}

export interface DatabaseAdapter
	extends BaseObserverInterface<DBAdapterListener>,
		DatabaseGetUtilities {
	close: () => void | Promise<void>;
	execute: (query: string, params?: any[]) => Promise<QueryResult>;
	executeRaw: (query: string, params?: any[]) => Promise<any[][]>;
	executeBatch: (query: string, params?: any[][]) => Promise<QueryResult>;
	name: string;
	readLock: <T>(
		fn: (tx: LockContext) => Promise<T>,
		options?: DatabaseLockOptions,
	) => Promise<T>;
	readTransaction: <T>(
		fn: (tx: Transaction) => Promise<T>,
		options?: DatabaseLockOptions,
	) => Promise<T>;
	writeLock: <T>(
		fn: (tx: LockContext) => Promise<T>,
		options?: DatabaseLockOptions,
	) => Promise<T>;
	writeTransaction: <T>(
		fn: (tx: Transaction) => Promise<T>,
		options?: DatabaseLockOptions,
	) => Promise<T>;
}